+++ module/User/src/User/Model/UserTable.php
@@ -7,20 +7,41 @@ class UserTable extends User
 {
 	protected $tableGateway;
 	protected $rolesTableGateway;
-
-	public function __construct(TableGateway $tableGateway, TableGateway $rolesTableGateway)
+	
+	public function __construct(TableGateway $tableGateway, TableGateway $rolesTableGateway = null)
 	{
 		$this->tableGateway = $tableGateway;
 		$this->rolesTableGateway = $rolesTableGateway;
 	}
 
+	/**
+	 * Retoune tous les utilisateurs existants
+	 * 
+	 * @return User[] Liste des utilisateurs (sous forme d'un iterable)
+	 */
 	public function fetchAll()
 	{
 		// TODO: Lors de la mise en place de la suppression logique, fitrer le select pour ne pas utiliser les comptes supprimés logiquement
 		$resultSet = $this->tableGateway->select();
 		return $resultSet;
 	}
-
+	
+	/**
+	 * Retourn tous les utilisateurs appartenant possedant un ID role donné
+	 * 
+	 * @param integer $roleId ID du rôle
+	 * @return User[] Liste des utilisateurs (sous forme d'un iterable)
+	 */
+	public function fetchByRole($roleId)
+	{
+		$users = array();
+		$roleSet = $this->rolesTableGateway->select(array('role_id' => $roleId));
+		foreach ($roleSet as $role) { 
+			$users[] = $this->getUser($role->user_id);
+		}
+		return $users;
+	}
+	
 	public function getUser($id)
 	{
 		$id  = (int) $id;
@@ -31,11 +52,13 @@ class UserTable extends User
 		}
 		
 		$roles = array();
-		$roleset = $this->rolesTableGateway->select(array(
-			'user_id' => $id
-		));
-		foreach ($roleset as $role) {
-			$roles[] = $role->role_id;
+		if ($this->rolesTableGateway) {
+			$roleset = $this->rolesTableGateway->select(array(
+				'user_id' => $id
+			));
+			foreach ($roleset as $role) {
+				$roles[] = $role->role_id;
+			}
 		}
 		$row->setRoles($roles);
 		
@@ -75,15 +98,17 @@ class UserTable extends User
 		}
 			
 		// Sauvegarde des droits du rôle
-		$this->rolesTableGateway->delete(array(
-			'user_id' => $id
-		));
-		$roles = $user->getRoles();
-		foreach ($roles as $role_id) {
-			$this->rolesTableGateway->insert(array(
-				'user_id' => $id,
-				'role_id' => $role_id
+		if ($this->rolesTableGateway) {
+			$this->rolesTableGateway->delete(array(
+				'user_id' => $id
 			));
+			$roles = $user->getRoles();
+			foreach ($roles as $role_id) {
+				$this->rolesTableGateway->insert(array(
+					'user_id' => $id,
+					'role_id' => $role_id
+				));
+			}
 		}
 	}
 
@@ -111,6 +136,9 @@ class UserTable extends User
 	
 	public function deleteUser($id)
 	{
+		if (!$this->rolesTableGateway) {
+			throw new \Exception("Can't delete an user, no rights TableGateway given.");
+		}
 		// Suppression des rôles de l'utilisateur
 		$this->rolesTableGateway->delete(array(
 			'user_id' => $id
@@ -121,7 +149,8 @@ class UserTable extends User
 		
 		// Suppression logique:
 		// Utiliser une colonne dédiée (bool) indiquant que le compte est supprimer logiquement.
-		// Modifier le champ email en del_ID_EMAIL où ID et l'id du compte, et EMAIL l'email actuel (pour éviter les problèmes d'unicité) 
+		// Modifier le champ email en del_ID_EMAIL où ID et l'id du compte, et EMAIL l'email actuel (pour éviter les problèmes d'unicité)
+		// Supprimer tous les rôles affecté au compte mis en suppression logique 
 		// Ne pas lister les comptes supprimés logiquement dans le fetchAll et autres
 	}
 }
